Make utility functions
I have made a util function for the
UI, so insted of setting up UI
components manually, you call a
pre-defined style.

I have also made a util function for
controller support. The InputUtil
must be instantiated, and its
update() function must be run on the
start of every render.